feat: balances yields speed
