+++ src/handlers/getLatestBalances.ts
@@ -1,11 +1,10 @@
+import { selectLatestBalancesGroupsByFromAddress } from '@db/balances'
 import { connect } from '@db/clickhouse'
 import type { BalancesResponse } from '@handlers/getBalances'
-import { formatBalancesGroups } from '@handlers/getBalances'
 import { badRequest, serverError, success } from '@handlers/response'
 import { updateBalances } from '@handlers/updateBalances'
 import { BALANCE_UPDATE_THRESHOLD_SEC } from '@lib/balance'
 import { isHex } from '@lib/buf'
-import { aggregateYields } from '@lib/yields'
 import type { APIGatewayProxyHandler } from 'aws-lambda'
 
 export const handler: APIGatewayProxyHandler = async (event, _context) => {
@@ -23,18 +22,14 @@ export const handler: APIGatewayProxyHandler = async (event, _context) => {
   const client = connect()
 
   try {
-    const balancesGroups = await updateBalances(client, address)
+    const { timestamp } = await updateBalances(client, address)
 
-    const updatedAt = balancesGroups[0]?.timestamp ? new Date(balancesGroups[0]?.timestamp).getTime() : undefined
-
-    const formattedBalancesGroups = formatBalancesGroups(balancesGroups)
-
-    await aggregateYields(formattedBalancesGroups)
+    const { updatedAt, balancesGroups } = await selectLatestBalancesGroupsByFromAddress(client, address, timestamp)
 
     const balancesResponse: BalancesResponse = {
       status: 'success',
-      updatedAt: updatedAt === undefined ? undefined : Math.floor(updatedAt / 1000),
-      groups: formattedBalancesGroups,
+      updatedAt,
+      groups: balancesGroups,
     }
 
     return success(balancesResponse, { maxAge: BALANCE_UPDATE_THRESHOLD_SEC })
